Move Debugger to own file and use type in Controller
